Sqlite works, made fixes in config so it becomes a basis, added a few mods on top
