+++ cognitive_architecture/database/create_database.py
@@ -70,66 +70,4 @@ class DatabaseManager:
         async with self.engine.begin() as conn:
             await conn.run_sync(Base.metadata.create_all)
 
-if __name__ == "__main__":
-    async def main():
-        """Runs as a part of startup docker scripts to create the database and tables."""
 
-        dbconfig = DatabaseConfig(db_type=config.db_type, db_name=config.db_name)
-        db_manager = DatabaseManager(config=dbconfig)
-        database_name = dbconfig.db_name
-
-        if not await db_manager.database_exists(database_name):
-            print(f"Database {database_name} does not exist. Creating...")
-            await db_manager.create_database(database_name)
-            print(f"Database {database_name} created successfully.")
-
-        await db_manager.create_tables()
-
-    asyncio.run(main())
-#
-# def create_admin_engine(username, password, host, database_name):
-#     admin_url = f"postgresql://{username}:{password}@{host}:5432/{database_name}"
-#     return create_engine(admin_url)
-#
-# def database_exists(connection, db_name):
-#     query = text(f"SELECT 1 FROM pg_database WHERE datname='{db_name}'")
-#     result = connection.execute(query).fetchone()
-#     return result is not None
-#
-# def create_database(connection, db_name):
-#     connection.set_isolation_level(psycopg2.extensions.ISOLATION_LEVEL_AUTOCOMMIT)
-#     cursor = connection.cursor()
-#     cursor.execute(f"CREATE DATABASE {db_name}")
-#     cursor.close()
-#
-# def drop_database(connection, db_name):
-#     connection.set_isolation_level(psycopg2.extensions.ISOLATION_LEVEL_AUTOCOMMIT)
-#     cursor = connection.cursor()
-#     cursor.execute(f"DROP DATABASE IF EXISTS {db_name}")
-#     cursor.close()
-#
-#
-#
-# def create_tables(engine):
-#     Base.metadata.create_all(bind = engine)
-#
-# if __name__ == "__main__":
-#     host = os.environ.get('POSTGRES_HOST')
-#     username = os.environ.get('POSTGRES_USER')
-#     password = os.environ.get('POSTGRES_PASSWORD')
-#     database_name = os.environ.get('POSTGRES_DB')
-#
-#     engine = create_admin_engine(username, password, host, database_name)
-#     connection = engine.connect()
-#
-#     # print(Base.metadata.tables)
-#
-#     if not database_exists(connection, database_name):
-#         logger.info(f"Database {database_name} does not exist. Creating...")
-#         create_database(connection, database_name)
-#         logger.info(f"Database {database_name} created successfully.")
-#
-#     connection.close()
-#     engine.dispose()
-#
-#     create_tables(engine)
